test/core: Fix most compile warnings reported by clang
Still remain warning like this

In file included from main.c:10:
../ctest.h:335:86: warning: format specifies type 'unsigned int' but the argument has type 'uintmax_t' (aka 'unsigned long long') [-Wformat]
  335 |         CTEST_ERR("%s:%d  expected %" PRIuMAX " bytes, got %" PRIuMAX, caller, line, (uintmax_t) expsize, (uintmax_t) realsize);
      |                                    ~~~~~~~~~~                                        ^~~~~~~~~~~~~~~~~~~
../ctest.h:335:107: warning: format specifies type 'unsigned int' but the argument has type 'uintmax_t' (aka 'unsigned long long') [-Wformat]
  335 |         CTEST_ERR("%s:%d  expected %" PRIuMAX " bytes, got %" PRIuMAX, caller, line, (uintmax_t) expsize, (uintmax_t) realsize);
      |                                                            ~~~~~~~~~~                                     ^~~~~~~~~~~~~~~~~~~~
../ctest.h:340:39: warning: format specifies type 'unsigned int' but the argument has type 'uintmax_t' (aka 'unsigned long long') [-Wformat]
  339 |             CTEST_ERR("%s:%d expected 0x%02x at offset %" PRIuMAX " got 0x%02x",
      |                                                        ~~~~~~~~~~
  340 |                 caller, line, exp[i], (uintmax_t) i, real[i]);
      |                                       ^~~~~~~~~~~~~
../ctest.h:347:80: warning: format specifies type 'int' but the argument has type 'intmax_t' (aka 'long long') [-Wformat]
  347 |         CTEST_ERR("%s:%d  expected %" PRIdMAX ", got %" PRIdMAX, caller, line, exp, real);
      |                                    ~~~~~~~~~~                                  ^~~
../ctest.h:347:85: warning: format specifies type 'int' but the argument has type 'intmax_t' (aka 'long long') [-Wformat]
  347 |         CTEST_ERR("%s:%d  expected %" PRIdMAX ", got %" PRIdMAX, caller, line, exp, real);
      |                                                      ~~~~~~~~~~                     ^~~~
../ctest.h:353:80: warning: format specifies type 'unsigned int' but the argument has type 'uintmax_t' (aka 'unsigned long long') [-Wformat]
  353 |         CTEST_ERR("%s:%d  expected %" PRIuMAX ", got %" PRIuMAX, caller, line, exp, real);
      |                                    ~~~~~~~~~~                                  ^~~
../ctest.h:353:85: warning: format specifies type 'unsigned int' but the argument has type 'uintmax_t' (aka 'unsigned long long') [-Wformat]
  353 |         CTEST_ERR("%s:%d  expected %" PRIuMAX ", got %" PRIuMAX, caller, line, exp, real);
      |                                                      ~~~~~~~~~~                     ^~~~
../ctest.h:359:67: warning: format specifies type 'int' but the argument has type 'intmax_t' (aka 'long long') [-Wformat]
  359 |         CTEST_ERR("%s:%d  should not be %" PRIdMAX, caller, line, real);
      |                                         ~~~~~~~~~~                ^~~~
../ctest.h:365:67: warning: format specifies type 'unsigned int' but the argument has type 'uintmax_t' (aka 'unsigned long long') [-Wformat]
  365 |         CTEST_ERR("%s:%d  should not be %" PRIuMAX, caller, line, real);
      |                                         ~~~~~~~~~~                ^~~~
../ctest.h:371:93: warning: format specifies type 'int' but the argument has type 'intmax_t' (aka 'long long') [-Wformat]
  371 |         CTEST_ERR("%s:%d  expected %" PRIdMAX "-%" PRIdMAX ", got %" PRIdMAX, caller, line, exp1, exp2, real);
      |                                    ~~~~~~~~~~                                               ^~~~
../ctest.h:371:99: warning: format specifies type 'int' but the argument has type 'intmax_t' (aka 'long long') [-Wformat]
  371 |         CTEST_ERR("%s:%d  expected %" PRIdMAX "-%" PRIdMAX ", got %" PRIdMAX, caller, line, exp1, exp2, real);
      |                                                 ~~~~~~~~~~                                        ^~~~
../ctest.h:371:105: warning: format specifies type 'int' but the argument has type 'intmax_t' (aka 'long long') [-Wformat]
  371 |         CTEST_ERR("%s:%d  expected %" PRIdMAX "-%" PRIdMAX ", got %" PRIdMAX, caller, line, exp1, exp2, real);
      |                                                                   ~~~~~~~~~~                            ^~~~
../ctest.h:533:34: warning: passing arguments to a function without a prototype is deprecated in all versions of C and is not supported in C2x [-Wdeprecated-non-prototype]
  533 |                         test->run(test->data);
      |                                  ^
13 warnings generated.

+++ test/core/test_bench.c
@@ -30,7 +30,7 @@ CTEST(bench, bench)
     }
     BENCH_STOP(memsetloop);
     BENCH_STAT(memsetloop);
-    printf("usecyc:%lu, lpcnt:%lu, sumcyc:%lu\n", (uint32_t)BENCH_GET_USECYC(), (uint32_t)BENCH_GET_LPCNT(), (uint32_t)BENCH_GET_SUMCYC());
+    printf("usecyc:%lu, lpcnt:%lu, sumcyc:%lu\n", (unsigned long)BENCH_GET_USECYC(), (unsigned long)BENCH_GET_LPCNT(), (unsigned long)BENCH_GET_SUMCYC());
 
 }
 // Declare HPMCOUNTER3 and HPMCOUNTER4
@@ -74,7 +74,7 @@ CTEST(bench, hpm)
 
     HPM_STAT(3, memsetloop, HPM_EVENT3);
     HPM_STAT(4, memsetloop, HPM_EVENT4);
-    printf("hpm3, usecyc:%lu, lpcnt:%lu, sumcyc:%lu\n", (uint32_t)HPM_GET_USECYC(3), (uint32_t)HPM_GET_LPCNT(3), (uint32_t)HPM_GET_SUMCYC(3));
-    printf("hpm4, usecyc:%lu, lpcnt:%lu, sumcyc:%lu\n", (uint32_t)HPM_GET_USECYC(4), (uint32_t)HPM_GET_LPCNT(4), (uint32_t)HPM_GET_SUMCYC(4));
+    printf("hpm3, usecyc:%lu, lpcnt:%lu, sumcyc:%lu\n", (unsigned long)HPM_GET_USECYC(3), (unsigned long)HPM_GET_LPCNT(3), (unsigned long)HPM_GET_SUMCYC(3));
+    printf("hpm4, usecyc:%lu, lpcnt:%lu, sumcyc:%lu\n", (unsigned long)HPM_GET_USECYC(4), (unsigned long)HPM_GET_LPCNT(4), (unsigned long)HPM_GET_SUMCYC(4));
 }
